Update Api & Controller

+++ app/Http/Controllers/Backend/UserController.php
@@ -8,6 +8,7 @@ use App\Models\User;
 use App\Models\Bidang;
 use App\Models\Jabatan;
 use App\Models\Pangkat;
+use App\Models\OpdChange;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -35,13 +36,22 @@ class UserController extends \App\Http\Controllers\Controller
             ->when($request->has('nama'), function ($query) use ($request) {
                 $query->where('name', 'like', '%' . $request->nama . '%');
             })
+            ->when($request->has('opd'), function ($query) use ($request) {
+                $query->whereHas('opd', function ($q) use ($request) {
+                    $q->where('name', 'like', '%' . $request->opd . '%');
+                });
+            })
             ->when($request->has('nip'), function ($query) use ($request) {
                 $query->where('nip', 'like', '%' . $request->nip . '%');
             })
+            ->orderByRaw('eselon_id IS NULL, eselon_id, pangkat_id IS NULL, pangkat_id, status = "PNS" DESC, status')
             ->paginate(10);
+        // Menambahkan query parameter ke pagination links
+        $users->appends($request->except('page'));
 
         return view('backend.pegawai.list_pegawai', compact('users'));
     }
+
 
 
     public function create()
@@ -60,12 +70,16 @@ class UserController extends \App\Http\Controllers\Controller
             'name' => 'required',
             'eselon_id' => 'nullable',
             'password' => 'required|min:8',
-            'nip' => 'nullable',
+            'nip' => 'nullable|unique:users', // Menambahkan aturan unique untuk NIP dengan pesan kesalahan khusus
             'status' => 'required',
             'opd_id' => 'nullable',
             'jabatan_id' => 'nullable',
             'pangkat_id' => 'nullable',
+        ], [
+            'password.min' => 'Panjang password minimal harus 8 karakter.',
+            'nip.unique' => 'NIP sudah ada dalam database. Harap masukkan NIP yang berbeda.',
         ]);
+
 
         $password = $validatedData['password'];
         $hashedPassword = Hash::make($password);
@@ -79,11 +93,17 @@ class UserController extends \App\Http\Controllers\Controller
         $user->opd_id = $validatedData['opd_id'];
         $user->jabatan_id = $validatedData['jabatan_id'];
         $user->pangkat_id = $validatedData['pangkat_id'];
-        $user->save();
+
+        try {
+            $user->save();
+        } catch (\Exception $e) {
+            return redirect()->route('user.index')->with('error', 'NIP sudah ada dalam database!');
+        }
 
         return redirect()->route('user.index')
             ->with('success', 'Anggota berhasil ditambahkan');
     }
+
 
     public function update(Request $request, $id)
     {
@@ -98,13 +118,11 @@ class UserController extends \App\Http\Controllers\Controller
             'role' => 'required',
             'nip' => 'nullable',
             'status' => 'nullable',
-            'nohp' => 'nullable',
             'opd_id' => 'nullable',
             'bidang_id' => 'nullable',
             'jabatan_id' => 'nullable',
             'pangkat_id' => 'nullable',
             'akses' => 'required',
-            'laporan_akses' => 'nullable',
             'eselon_id' => 'nullable',
         ]);
 
@@ -113,13 +131,25 @@ class UserController extends \App\Http\Controllers\Controller
         $user->role = $validatedData['role'];
         $user->nip = $validatedData['nip'];
         $user->status = $validatedData['status'];
-        $user->nohp = $validatedData['nohp'];
         $user->opd_id = $validatedData['opd_id'];
         $user->bidang_id = $validatedData['bidang_id'];
         $user->jabatan_id = $validatedData['jabatan_id'];
         $user->pangkat_id = $validatedData['pangkat_id'];
         $user->akses = $validatedData['akses'];
-        $user->laporan_akses = $validatedData['laporan_akses'];
+
+        // Check if OPD has changed
+        if ($user->isDirty('opd_id')) {
+            $oldOpdId = $user->getOriginal('opd_id');
+            $newOpdId = $user->opd_id;
+
+            // Create OpdChange record
+            OpdChange::create([
+                'user_id' => $user->id,
+                'old_opd_id' => $oldOpdId,
+                'new_opd_id' => $newOpdId,
+                'tanggal_pindah' => now(), // Tanggal pindah diisi dengan waktu saat ini
+            ]);
+        }
 
         $update = $user->save();
 
@@ -134,9 +164,19 @@ class UserController extends \App\Http\Controllers\Controller
     public function destroy($id)
     {
         $user = User::find($id);
-        $user->delete();
-        return redirect()->route('user.index')->with('success', 'Anggota berhasil dihapus!');
+
+        if (!$user) {
+            return redirect()->route('user.index')->with('error', 'Anggota tidak ditemukan!');
+        }
+
+        try {
+            $user->delete();
+            return redirect()->route('user.index')->with('success', 'Anggota berhasil dihapus!');
+        } catch (\Exception $e) {
+            return redirect()->route('user.index')->with('error', 'Gagal menghapus anggota. Silakan coba lagi.');
+        }
     }
+
 
     public function resetPassword(Request $request, $id)
     {
